requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
scikit-image==0.22.0
numpy==1.26.2
pandas==2.1.3
matplotlib==3.8.2
Pillow==10.1.0
scipy==1.11.4
